# TOOLS/lua/autoload: avoid unnecessary playlist manipulation, performance