We used to sort the playlist with playlist-move after every loadfile (slow operation).  
Instead, append all files in order and call playlist-move once to move the only entry we don't control the position of.  
Don't fetch every playlist item separately, reuse native property.  
We used to pick up on new files added to the directory, but only when playing an entry at the edge of the playlist due to an early return.  
New files are now added to the playlist on every file change.  
This still works as expected and doesn't load duplicate files on shuffled playlists or playlists with files manually added after autoload.

33% faster on average for my test directory with 1371 files. Initial run of `find_and_add_entries` went from 1.6s-2.2s (2s avg) to 1.1s-1.4s (1.3s avg).

Not moving the current playlist entry's position many times should also mitigate race conditions in other scripts that interact with playlists.