@po5 2023-07-02
Note that manually removed (`playlist-remove n`) autoloaded entries will be re-added when the next file is loaded. This was also an issue with the old code, but it only manifested itself when the removed entry was beyond the bail check (https://github.com/mpv-player/mpv/pull/11856#discussion_r1248929369 for details).  
Shouldn't be a blocker imo, but should definitely be addressed at some point. I think the best way is to make the `filenames` table persist across file loads within the same directory. Known entries will not be re-added.