@christoph-heinrich 2023-07-02
#11292 keeps track of added files to avoid re-adding them.